carrito100%funcional

+++ Paginaweb2/core/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class Producto(models.Model):
     codigo = models.CharField(max_length=10)
     nombre = models.CharField(max_length=100)
-    precio = models.DecimalField(max_digits=8, decimal_places=2)
+    precio = models.IntegerField()
     descripcion = models.TextField()
 
     def __str__(self):
